@@ -10,37 +10,36 @@
 		"check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch"
 	},
 	"devDependencies": {
-		"@sveltejs/adapter-auto": "^3.0.0",
-		"@sveltejs/kit": "^2.0.0",
-		"@sveltejs/vite-plugin-svelte": "^3.0.0",
-		"@tailwindcss/typography": "^0.5.13",
+		"@sveltejs/adapter-auto": "^4.0.0",
+		"@sveltejs/kit": "^2.17.1",
+		"@sveltejs/vite-plugin-svelte": "^5.0.3",
+		"@tailwindcss/typography": "^0.5.14",
 		"@types/better-sqlite3": "^7.6.12",
-		"@types/jsonwebtoken": "^9.0.6",
-		"@types/three": "^0.167.2",
-		"autoprefixer": "^10.4.19",
-		"flowbite": "^2.4.1",
-		"flowbite-svelte": "^0.46.15",
-		"postcss": "^8.4.38",
-		"svelte": "^4.2.7",
-		"svelte-check": "^3.6.0",
-		"tailwindcss": "^3.4.4",
-		"typescript": "^5.0.0",
-		"vite": "^5.0.3"
+		"@types/jsonwebtoken": "^9.0.8",
+		"@types/three": "^0.173.0",
+		"autoprefixer": "^10.4.20",
+		"flowbite": "^3.1.1",
+		"flowbite-svelte": "^0.47.4",
+		"svelte": "^5.19.7",
+		"svelte-check": "^4.1.4",
+		"tailwindcss": "^3.4.9",
+		"typescript": "^5.7.3",
+		"vite": "^6.0.11"
 	},
 	"type": "module",
 	"dependencies": {
-		"@discordjs/builders": "^1.8.2",
+		"@discordjs/builders": "^1.10.0",
 		"@oslojs/crypto": "^1.0.1",
 		"@oslojs/encoding": "^1.1.0",
-		"@sveltejs/adapter-node": "^5.2.1",
-		"arctic": "^3.2.1",
+		"@sveltejs/adapter-node": "^5.2.12",
+		"arctic": "^3.2.2",
 		"better-sqlite3": "^11.8.1",
 		"dayjs": "^1.11.13",
-		"discord.js": "^14.15.3",
-		"dotenv": "^16.4.5",
+		"discord.js": "^14.17.3",
+		"dotenv": "^16.4.7",
 		"jsonwebtoken": "^9.0.2",
 		"short-unique-id": "^5.2.0",
 		"skinview-utils": "^0.7.1",
-		"three": "^0.167.1"
+		"three": "^0.173.0"
 	}
 }
