@@ -63,8 +63,8 @@
 
 <div class="w-full flex-1 p-2">
     <div class="flex w-full mb-5 rounded-md">
-        <div class="w-2/6 mr-5">
-        <PanelLeft data={data.datas} {viewer} bind:getSkinDatas={getSkinDatas} bind:loadDefault={ldSkinFn}/>
+        <div class="w-2/6 mr-5 max-h-[65vh]">
+        <PanelLeft data={data.datas} {viewer} bind:getSkinData={getSkinDatas} bind:loadDefault={ldSkinFn}/>
         </div>
         <div class="rounded-md bgimg flex grow justify-center">
             <Viewer bind:viewer={viewer} bind:createSkin={exportFn} />
